Add files via upload
A vending machine using FPGAs and FSM is developed using Xilinx and Quartus Prime. The machine will dispense three products A, B, and C which costs Rs.10, Rs.15, and Rs.20 respectively. It will also show the remaining change to be returned when a higher denomination coin is inserted. This is verified with the help of the output waveforms.